refactor: extract ManagePage save/submit into focused hooks
Split draft autosave and publish flows out of ManagePage into
per-domain hooks so knowledge/tools/LLM changes live in dedicated
files without changing manage-page behavior (#432).
