+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "(.*pnpm-lock.*|.*js.*|node_modules|.venv|.*jinja2.*|.*woff2.*)|^.secrets.baseline$|^.env$",
     "lines": null
   },
-  "generated_at": "2026-07-06T20:41:10Z",
+  "generated_at": "2026-07-09T12:07:34Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -684,7 +684,7 @@
         "hashed_secret": "1f5e25be9b575e9f5d39c82dfd1d9f4d73f1975c",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 230,
+        "line_number": 250,
         "type": "Secret Keyword",
         "verified_result": null
       },
@@ -692,7 +692,7 @@
         "hashed_secret": "d0a3e7f81a9885e99049d1cae0336d269d5e47a9",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 283,
+        "line_number": 292,
         "type": "Secret Keyword",
         "verified_result": null
       },
@@ -700,7 +700,7 @@
         "hashed_secret": "48bf2239eea43073d82dc9a4ad45200b8e297363",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 461,
+        "line_number": 464,
         "type": "Secret Keyword",
         "verified_result": null
       },
@@ -708,7 +708,7 @@
         "hashed_secret": "6947818ac409551f11fbaa78f0ea6391960aa5b8",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 462,
+        "line_number": 465,
         "type": "Secret Keyword",
         "verified_result": null
       }
