+++ src/frontend_workspaces/frontend/tsconfig.json
@@ -12,5 +12,6 @@
     "moduleResolution": "node",
     "preserveConstEnums": true
   },
-  "include": ["src/**/*"]
+  "include": ["src/**/*"],
+  "exclude": ["src/**/*.test.ts", "src/**/*.test.tsx"]
 }
